Argument & RSP support file segmentation
